GeometricEquations

Documentation for GeometricEquations.

In GeometricEquations.jl we define three basic types of equations:

ordinary differential equations (ODEs),

differential algebraic equations (DAEs),

stochastic differential equations (SDEs).

For each type, there are several subtypes

standard equations (ODE, DAE, SDE),

implicit equations (IODE, IDAE),

partitioned equations (PODE, PDAE, PSDE),

Hamiltonian equations (HODE, HDAE),

Lagrangian equations (LODE, LDAE),

split equations (SODE, SPDAE), SPSDE).

Each equation holds a number of functions determining the vector field, constraints, initial conditions, and possibly additional information like parameters, periodicity, invariants and the Hamiltonian or Lagrangian.

Library

GeometricEquations.DAE

GeometricEquations.DAEProblem

GeometricEquations.GeometricProblem

GeometricEquations.HDAE

GeometricEquations.HDAEProblem

GeometricEquations.HODE

GeometricEquations.HODEProblem

GeometricEquations.IDAE

GeometricEquations.IDAEProblem

GeometricEquations.IODE

GeometricEquations.IODEProblem

GeometricEquations.LDAE

GeometricEquations.LDAEProblem

GeometricEquations.LODE

GeometricEquations.LODEProblem

GeometricEquations.ODE

GeometricEquations.ODEProblem

GeometricEquations.PDAE

GeometricEquations.PDAEProblem

GeometricEquations.PODE

GeometricEquations.PODEProblem

GeometricEquations.PSDE

GeometricEquations.PSDEProblem

GeometricEquations.SDE

GeometricEquations.SDEProblem

GeometricEquations.SODE

GeometricEquations.SODEProblem

GeometricEquations.SPDAE

GeometricEquations.SPDAEProblem

GeometricEquations.SPSDE

GeometricEquations.SPSDEProblem

GeometricBase.equation

GeometricBase.functions

GeometricBase.invariants

GeometricBase.solutions

GeometricEquations.promote_tspan

GeometricEquations.DAE — Type

DAE: Differential Algebraic Equation

Defines a differential algebraic initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t, q(t)) + u(t, q(t), \lambda(t)) , & q(t_{0}) &= q_{0} , \\ 0 &= \phi (t, q(t), \lambda(t)) , & \lambda(t_{0}) &= \lambda_{0} , \end{aligned}\]

with vector field $v$, projection $u$, algebraic constraint $\phi=0$, initial conditions $q_{0}$ and $\lambda_{0}$, the dynamical variable $q$ taking values in $\mathbb{R}^{d}$ and the algebraic variable $\lambda$ taking values in $\mathbb{R}^{m}$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

vType <: Function: type of v

uType <: Function: type of u

ūType <: OptionalFunction: type of ū

ϕType <: Function: type of ϕ

ψType <: OptionalFunction: type of ψ

v̄Type <: Function: type of v̄

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

d: dimension of dynamical variable $q$ and the vector field $v$

m: dimension of algebraic variable $\lambda$ and the constraint $\phi$

v: function computing the vector field

u: function computing the projection

ū: function computing the secondary projection field $\bar{u}$ (optional)

ϕ: algebraic constraint

ψ: secondary constraints (optional)

v̄: function computing an initial guess for the velocity field $v$ (defaults to v)

t₀: initial time

q₀: initial condition for dynamical variable $q$

λ₀: initial condition for algebraic variable $\lambda$

μ₀: initial condition for algebraic variable $μ$ (optional)

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

The function v, providing the vector field, takes three arguments, v(t, q, v), the functions u and ϕ, providing the projection and the algebraic constraint take four arguments, u(t, q, λ, u) and ϕ(t, q, λ, ϕ), where t is the current time, q and λ are the current solution vectors, and v, u and ϕ are the vectors which hold the result of evaluating the vector field $v$, the projection $u$ and the algebraic constraint $\phi$ on t, q and λ.

Constructors

DAE(v, u, ū, ϕ, ψ, v̄, t₀, q₀, λ₀, invariants, parameters, periodicity)

DAE(v, u, ϕ, t₀, q₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
DAE(v, u, ϕ, q₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
DAE(v, u, ϕ, t₀, q₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
DAE(v, u, ϕ, q₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)

DAE(v, u, ū, ϕ, ψ, t₀, q₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
DAE(v, u, ū, ϕ, ψ, q₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
DAE(v, u, ū, ϕ, ψ, t₀, q₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
DAE(v, u, ū, ϕ, ψ, q₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)

Example

    function v(t, q, v)
        v[1] = q[1]
        v[2] = q[2]
    end

    function u(t, q, λ, u)
        u[1] = +λ[1]
        u[2] = -λ[1]
    end

    function ϕ(t, q, λ, ϕ)
        ϕ[1] = q[2] - q[1]
    end

    t₀ = 0.
    q₀ = [1., 1.]
    λ₀ = [0.]

    dae = DAE(v, u, ϕ, t₀, q₀, λ₀)

GeometricEquations.DAEProblem — Type

DAEProblem: Differential Algebraic Equation Problem

Defines a differential algebraic initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t, q(t)) + u(t, q(t), \lambda(t)) , & q(t_{0}) &= q_{0} , \\ 0 &= \phi (t, q(t), \lambda(t)) , & \lambda(t_{0}) &= \lambda_{0} , \end{aligned}\]

with vector field $v$, projection $u$, algebraic constraint $\phi=0$, and initial conditions $q_{0}$ and $\lambda_{0}$.

GeometricEquations.GeometricProblem — Type

GeometricProblem: stores a GeometricEquation togehter with initial conditions, parameters, time span and time step size.

Parameters

ST <: GeometricEquation: super type, used for dispatch

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

equType <: GeometricEquation: equation type

icsType <: paramsType: initial conditions type

parType <: OptionalParameters: parameters type

Fields

equation: reference to the parent equation object

tspan: time span for problem (t₀,t₁)

tstep: time step to be used in simulation

ics: NamedTuple containing the initial conditions

parameters: either a NamedTuple containing the equation's parameters or NullParameters

Subtypes

The GeometricProblem type has various subtypes for the different equations types, that are defined e.g. via

const ODEProblem = GeometricProblem{ODE}

and provide convenience constructors to construct and equation and the corresponding problem in one step.

GeometricEquations.HDAE — Type

HDAE: Hamiltonian Differential Algebraic Equation

Defines a Hamiltonian differential algebraic initial value problem, that is a canonical Hamiltonian system of equations subject to Dirac constraints,

\[\begin{aligned} \dot{q} (t) &= v(t, q(t), p(t)) + u(t, q(t), p(t), \lambda(t)) + \bar{g}(t, q(t), p(t), \lambda(t), \gamma(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) + g(t, q(t), p(t), \lambda(t)) + \bar{f}(t, q(t), p(t), \lambda(t), \gamma(t)) , & p(t_{0}) &= p_{0} , \\ 0 &= \phi (t, q(t), p(t)) , \\ 0 &= \psi (t, q(t), p(t), \dot{q}(t), \dot{p}(t)) , \end{aligned}\]

with vector fields $v$, $u$, $\bar{u}$ and $f$, $g$, $\bar{g}$, primary constraint $\phi(q,p)=0$ and secondary constraint $\psi(q,p,\lambda)=0$, initial conditions $(q_{0}, p_{0})$, the dynamical variables $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$ and the algebraic variables $(\lambda, \gamma)$ taking values in $\mathbb{R}^{m} \times \mathbb{R}^{m}$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

vType <: Function: type of v

fType <: Function: type of f

uType <: Function: type of u

gType <: Function: type of g

ϕType <: Function: type of ϕ

ūType <: Function: type of ū

ḡType <: Function: type of ḡ

ψType <: Function: type of ψ

PType <: Function: type of P

v̄Type <: Function: type of v̄

f̄Type <: Function: type of f̄

hamType <: Function: Hamiltonian type

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $v$ and $f$

m: dimension of algebraic variables $\lambda$ and $\gamma$ and the constraints $\phi$ and $\psi$

v: function computing the Hamiltonian vector field $v$

f: function computing the Hamiltonian vector field $f$

u: function computing the projection for $q$

g: function computing the primary projection field $g$

ϕ: primary constraints

ū: function computing the secondary projection field $\bar{u}$ (optional)

ḡ: function computing the secondary projection field $\bar{g}$ (optional)

ψ: secondary constraints (optional)

P: function computing the Poisson matrix $P$

v̄: function computing an initial guess for the velocity field $v$(optional, defaults tov`)

f̄: function computing an initial guess for the force field $f$ (optional, defaults to f)

t₀: initial time (optional)

q₀: initial condition for dynamical variable $q$

p₀: initial condition for dynamical variable $p$

λ₀: initial condition for algebraic variable $λ$

μ₀: initial condition for algebraic variable $μ$ (optional)

hamiltonian: function computing the Hamiltonian $H$

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

Constructors

HDAE(v, f, u, g, ϕ, ū, ḡ, ψ, v̄, f̄, P, t₀, q₀, p₀, λ₀, hamiltonian, invariants, parameters, periodicity)

HDAE(v, f, u, g, ϕ, h, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...)
HDAE(v, f, u, g, ϕ, h, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...)
HDAE(v, f, u, g, ϕ, h, t₀, q₀::State, p₀::State, λ₀::State; kwargs...)
HDAE(v, f, u, g, ϕ, h, q₀::State, p₀::State, λ₀::State; kwargs...)

HDAE(v, f, u, g, ϕ, ū, ḡ, ψ, h, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...)
HDAE(v, f, u, g, ϕ, ū, ḡ, ψ, h, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...)
HDAE(v, f, u, g, ϕ, ū, ḡ, ψ, h, t₀, q₀::State, p₀::State, λ₀::State; kwargs...)
HDAE(v, f, u, g, ϕ, ū, ḡ, ψ, h, q₀::State, p₀::State, λ₀::State; kwargs...)

GeometricEquations.HDAEProblem — Type

HDAE: Hamiltonian Differential Algebraic Equation

Defines a Hamiltonian differential algebraic initial value problem, that is a canonical Hamiltonian system of equations subject to Dirac constraints,

\[\begin{aligned} \dot{q} (t) &= v(t, q(t), p(t)) + u(t, q(t), p(t), \lambda(t)) + \bar{g}(t, q(t), p(t), \lambda(t), \gamma(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) + g(t, q(t), p(t), \lambda(t)) + \bar{f}(t, q(t), p(t), \lambda(t), \gamma(t)) , & p(t_{0}) &= p_{0} , \\ 0 &= \phi (t, q(t), p(t)) , \\ 0 &= \psi (t, q(t), p(t), \dot{q}(t), \dot{p}(t)) , \end{aligned}\]

with vector fields $v$, $u$, $\bar{u}$ and $f$, $g$, $\bar{g}$, primary constraint $\phi(q,p)=0$ and secondary constraint $\psi(q,p,\lambda)=0$, and initial conditions $(q_{0}, p_{0})$.

GeometricEquations.HODE — Type

HODE: Hamiltonian Ordinary Differential Equation

Defines a Hamiltonian ordinary differential initial value problem, that is a canonical Hamiltonian system of equations,

\[\begin{aligned} \dot{q} (t) &= v(t, q(t), p(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) , & p(t_{0}) &= p_{0} , \end{aligned}\]

with vector fields $v$ and $f$, given by

\[\begin{aligned} v &= \frac{\partial H}{\partial p} , & f &= - \frac{\partial H}{\partial q} , \end{aligned}\]

initial conditions $(q_{0}, p_{0})$ and the dynamical variables $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

vType <: Function: type of v

fType <: Function: type of f

PType <: Function: type of P

hamType <: Function: Hamiltonian type

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $v$ and $f$

v: function computing the vector field $v$

f: function computing the vector field $f$

poisson: function computing the Poisson matrix $P$

hamiltonian: function computing the Hamiltonian $H$

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

Constructors

HODE(v, f, poisson, t₀, q₀, p₀, hamiltonian, invariants, parameters, periodicity)

HODE(v, f, h, t₀, q₀::StateVector, p₀::StateVector; kwargs...)
HODE(v, f, h, q₀::StateVector, p₀::StateVector; kwargs...)
HODE(v, f, h, t₀, q₀::State, p₀::State; kwargs...)
HODE(v, f, h, q₀::State, p₀::State; kwargs...)

Keyword arguments:

poisson = symplectic_matrix

invariants = nothing

parameters = nothing

periodicity = nothing

GeometricEquations.HODEProblem — Type

HODEProblem: Hamiltonian Ordinary Differential Equation Problem

Defines a Hamiltonian ordinary differential initial value problem, that is a canonical Hamiltonian system of equations,

\[\begin{aligned} \dot{q} (t) &= v(t, q(t), p(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) , & p(t_{0}) &= p_{0} , \end{aligned}\]

with vector fields $v$ and $f$, given by

\[\begin{aligned} v &= \frac{\partial H}{\partial p} , & f &= - \frac{\partial H}{\partial q} , \end{aligned}\]

initial conditions $(q_{0}, p_{0})$ and the dynamical variables $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$.

Constructors

HODEProblem(v, f, poisson, hamiltonian, tspan, tstep, ics; kwargs...)
HODEProblem(v, f, poisson, hamiltonian, tspan, tstep, q₀::State, p₀::State; kwargs...)

where ics is a NamedTuple with entries q and p.

Keyword arguments:

invariants = NullInvariants()

parameters = NullParameters()

periodicity = NullPeriodicity()

GeometricEquations.IDAE — Type

IDAE: Implicit Differential Algebraic Equation

Defines an implicit differential algebraic initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t) + u(t, q(t), p(t), \lambda(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) + g(t, q(t), p(t), \lambda(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= p(t, q(t), v(t)) , && \\ 0 &= \phi (t, q(t), p(t)) , & \lambda(t_{0}) &= \lambda_{0} , \end{aligned}\]

with force field $f$, the momentum defined by $p$, projection $g$ and $r$, algebraic constraint $\phi=0$, conditions $(q_{0}, p_{0})$ and $\lambda_{0}$, the dynamical variables $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$ and the algebraic variable $\lambda$ taking values in $\mathbb{R}^{m}$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

ϑType <: Function: type of ϑ

fType <: Function: type of f

uType <: Function: type of u

gType <: Function: type of g

ϕType <: Function: type of ϕ

ūType <: Function: type of ū

ḡType <: Function: type of ḡ

ψType <: Function: type of ψ

v̄Type <: Function: type of v̄

f̄Type <: Function: type of f̄

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $f$ and $p$

m: dimension of algebraic variable $\lambda$ and the constraint $\phi$

ϑ: function determining the momentum

f: function computing the vector field $f$

u: function computing the projection for $q$

g: function computing the projection for $p$

ϕ: algebraic constraints

ū: function computing the secondary projection field $\bar{u}$ (optional)

ḡ: function computing the secondary projection field $\bar{g}$ (optional)

ψ: secondary constraints (optional)

v̄: function computing an initial guess for the velocity field $v$ (optional)

f̄: function computing an initial guess for the force field $f$ (optional)

t₀: initial time (optional)

q₀: initial condition for dynamical variable $q$

p₀: initial condition for dynamical variable $p$

λ₀: initial condition for algebraic variable λ

μ₀: initial condition for algebraic variable μ (optional)

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

Constructors

IDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, v̄, f̄, t₀, q₀, p₀, λ₀, μ₀, invariants, parameters, periodicity)

IDAE(ϑ, f, u, g, ϕ, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
IDAE(ϑ, f, u, g, ϕ, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
IDAE(ϑ, f, u, g, ϕ, t₀, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
IDAE(ϑ, f, u, g, ϕ, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)

IDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
IDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
IDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, t₀, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
IDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)

Keyword arguments

v̄ = (t,q,v) -> nothing

f̄ = f

invariants = nothing

parameters = nothing

periodicity = nothing

GeometricEquations.IDAEProblem — Type

IDAEProblem: Implicit Differential Algebraic Equation Problem

Defines an implicit differential algebraic initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t) + u(t, q(t), p(t), \lambda(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) + g(t, q(t), p(t), \lambda(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= \vartheta(t, q(t), v(t)) , && \\ 0 &= \phi (t, q(t), p(t)) , & \lambda(t_{0}) &= \lambda_{0} , \end{aligned}\]

with force field $f$, the momentum defined by $p$, projection $u$ and $g$, algebraic constraint $\phi=0$, and initial conditions $(q_{0}, p_{0})$ and $\lambda_{0}$.

GeometricEquations.IODE — Type

IODE: Implicit Ordinary Differential Equation

Defines an implicit initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= ϑ(t, q(t), v(t)) \end{aligned}\]

with force field $f$, the momentum defined by $p$, initial conditions $(q_{0}, p_{0})$ and the solution $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$. This is a special case of a differential algebraic equation with dynamical variables $(q,p)$ and algebraic variable $v$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

ϑType <: Function: type of ϑ

fType <: Function: type of f

gType <: Function: type of g

v̄Type <: Function: type of v̄

f̄Type <: Function: type of f̄

hType <: OptionalFunction: type of h

pType <: Union{NamedTuple,Nothing}: parameters type

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $f$ and $p$

ϑ: function determining the momentum

f: function computing the vector field

g: function determining the projection, given by $\nabla \vartheta (q) \cdot \lambda$

v̄: function computing an initial guess for the velocity field $v$ (optional)

f̄: function computing an initial guess for the force field $f$ (optional)

h: function computing the Hamiltonian (optional)

t₀: initial time (optional)

q₀: initial condition for q

p₀: initial condition for p

parameters: either a NamedTuple containing the equations parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

The functions ϑ and f must have the interface

    function ϑ(t, q, v, p)
        p[1] = ...
        p[2] = ...
        ...
    end

and

    function f(t, q, v, f)
        f[1] = ...
        f[2] = ...
        ...
    end

where t is the current time, q is the current solution vector, v is the current velocity and f and p are the vectors which hold the result of evaluating the functions $f$ and $ϑ$ on t, q and v. In addition, the functions g, v̄ and f̄ are specified by

    function g(t, q, λ, g)
        g[1] = ...
        g[2] = ...
        ...
    end

    function v̄(t, q, v)
        v[1] = ...
        v[2] = ...
        ...
    end

    function f̄(t, q, v, f)
        f[1] = ...
        f[2] = ...
        ...
    end

The function g is used in projection methods that enforce $p = ϑ(q)$. The functions v̄ and f̄ are used for initial guesses in nonlinear implicit solvers.

Constructors

IODE(ϑ, f, v̄, f̄, t₀, q₀, p₀, λ₀, invariants, parameters, periodicity)

IODE(ϑ, f, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector=zero(q₀); kwargs...)
IODE(ϑ, f, q₀::StateVector, p₀::StateVector, λ₀::StateVector=zero(q₀); kwargs...)
IODE(ϑ, f, t₀, q₀::State, p₀::State, λ₀::StateVector=zero(q₀); kwargs...)
IODE(ϑ, f, q₀::State, p₀::State, λ₀::StateVector=zero(q₀); kwargs...)

Keyword arguments

v̄ = (t,q,v) -> nothing

f̄ = f

invariants = nothing

parameters = nothing

periodicity = nothing

GeometricEquations.IODEProblem — Type

IODEProblem: Implicit Ordinary Differential Equation Problem

Defines an implicit initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= ϑ(t, q(t), v(t)) \end{aligned}\]

with momentum defined by $\vartheta$, force field $f$, projection field $g$ and initial conditions $(q_{0}, p_{0}, \lambda_{0})$. This is a special case of a differential algebraic equation with dynamical variables $(q,p)$ and algebraic variable $v$

GeometricEquations.LDAE — Type

LDAE: Lagrangian Differential Algebraic Equation

Defines a Lagrangian differential algebraic initial value problem, that is a special implicit initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t) + \lambda(t), & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) + g(t, q(t), \lambda(t)) + \bar{g} (t, q(t), \mu(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= ϑ(t, q(t), v(t)) , \\ 0 &= \phi (t, q(t), p(t)) , \\ 0 &= \psi (t, q(t), p(t), \dot{q}(t), \dot{p}(t)) , \end{aligned}\]

with vector field $f$, the momentum defined by $p$, initial conditions $(q_{0}, p_{0})$ and the solution $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$ and the algebraic variables $(v, \lambda, \mu)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{m} \times \mathbb{R}^{m}$. This is a special case of a differential algebraic equation with dynamical variables $(q,p)$ and algebraic variables $v$, $\lambda$ and $\mu$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

ϑType <: Function: type of ϑ

fType <: Function: type of f

uType <: Function: type of u

gType <: Function: type of g

ϕType <: Function: type of ϕ

ūType <: Function: type of ū

ḡType <: Function: type of ḡ

ψType <: Function: type of ψ

ωType <: Function: type of ω

v̄Type <: Function: type of v̄

f̄Type <: Function: type of f̄

lagType <: Function: Lagrangian type

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $f$ and $p$

ϑ: function determining the momentum

f: function computing the vector field

u: function computing the projection for $q$, for a degenerate system given by $\lambda$

g: function computing the projection for $p$, for a degenerate system given by $\nabla \vartheta (q) \cdot \lambda$

ϕ: primary constraints, for a degenerate system given by $p - \vartheta (q)$

ū: function computing the secondary projection field $\bar{u}$, for a degenerate system given by $\lambda$ (optional)

ḡ: function computing the secondary projection field $\bar{g}$, for a degenerate system given by $\lambda \cdot \nabla \vartheta (q)$ (optional)

ψ: secondary constraints, for a degenerate system given by $\dot{p} - \dot{q} \cdot \nabla \vartheta (q)$ (optional)

ω: function computing the symplectic matrix

v̄: function computing an initial guess for the velocity field $v$ (optional)

f̄: function computing an initial guess for the force field $f$ (optional)

t₀: initial time (optional)

q₀: initial condition for dynamical variable q

p₀: initial condition for dynamical variable p

λ₀: initial condition for algebraic variable λ

μ₀: initial condition for algebraic variable μ (optional)

lagrangian: function computing the Lagrangian $L$

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

The functions ϑ and f must have the interface

    function ϑ(t, q, v, p)
        p[1] = ...
        p[2] = ...
        ...
    end

and

    function f(t, q, v, f)
        f[1] = ...
        f[2] = ...
        ...
    end

where t is the current time, q is the current solution vector, v is the current velocity and f and p are the vectors which hold the result of evaluating the functions $f$ and $ϑ$ on t, q and v. The funtions g, v̄ and f̄ are specified by

    function g(t, q, λ, g)
        g[1] = ...
        g[2] = ...
        ...
    end

    function v̄(t, q, v)
        v[1] = ...
        v[2] = ...
        ...
    end

    function f̄(t, q, v, f)
        f[1] = ...
        f[2] = ...
        ...
    end

Constructors

LDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, ω, v̄, f̄, t₀, q₀, p₀, λ₀, μ₀, lagrangian, invariants, parameters, periodicity)

LDAE(ϑ, f, u, g, ϕ, l, ω, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
LDAE(ϑ, f, u, g, ϕ, l, ω, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
LDAE(ϑ, f, u, g, ϕ, l, ω, t₀, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
LDAE(ϑ, f, u, g, ϕ, l, ω, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)

LDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, l, ω, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
LDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, l, ω, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
LDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, l, ω, t₀, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
LDAE(ϑ, f, u, g, ϕ, ū, ḡ, ψ, l, ω, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)

Keyword arguments

v̄ = (t,q,v) -> nothing

f̄ = f

invariants = nothing

parameters = nothing

periodicity = nothing

GeometricEquations.LDAEProblem — Type

LDAEProblem: Lagrangian Differential Algebraic Equation Problem

Defines a Lagrangian differential algebraic initial value problem, that is a special implicit initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t) + \lambda(t), & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) + g(t, q(t), \lambda(t)) + \bar{g} (t, q(t), \mu(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= ϑ(t, q(t), v(t)) , \\ 0 &= \phi (t, q(t), p(t)) , \\ 0 &= \psi (t, q(t), p(t), \dot{q}(t), \dot{p}(t)) , \end{aligned}\]

with vector field $f$, the momentum defined by $p$, and initial conditions $(q_{0}, p_{0}, \lambda_{0})$. This is a special case of a differential algebraic equation with dynamical variables $(q,p)$ and algebraic variables $v$, $\lambda$ and $\mu$.

GeometricEquations.LODE — Type

LODE: Lagrangian Ordinary Differential Equation

Defines an implicit initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= ϑ(t, q(t), v(t)) \end{aligned}\]

with momentum $p$ and force field $f$, given by

\[\begin{aligned} p &= \frac{\partial L}{\partial v} , & f &= \frac{\partial L}{\partial q} , \end{aligned}\]

initial conditions $(q_{0}, p_{0})$ and the solution $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$. This is a special case of a differential algebraic equation with dynamical variables $(q,p)$ and algebraic variable $v$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

ϑType <: Function: type of ϑ

fType <: Function: type of f

gType <: Function: type of g

ωType <: Function: type of ω

v̄Type <: Function: type of v̄

f̄Type <: Function: type of f̄

lagType <: Function: Lagrangian type

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $f$ and $p$

ϑ: function determining the momentum

f: function computing the vector field

g: function determining the projection, given by $\nabla \vartheta (q) \cdot \lambda$

ω: function computing the symplectic matrix

v̄: function computing an initial guess for the velocity field $v$ (optional)

f̄: function computing an initial guess for the force field $f$ (optional)

t₀: initial time (optional)

q₀: initial condition for q

p₀: initial condition for p

λ₀: initial condition for λ (optional)

lagrangian: function computing the Lagrangian $L$

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

The functions ϑ and f must have the interface

    function ϑ(t, q, v, p)
        p[1] = ...
        p[2] = ...
        ...
    end

and

    function f(t, q, v, f)
        f[1] = ...
        f[2] = ...
        ...
    end

where t is the current time, q is the current solution vector, v is the current velocity and f and p are the vectors which hold the result of evaluating the functions $f$ and $ϑ$ on t, q and v. The funtions g and v are specified by

    function g(t, q, λ, g)
        g[1] = ...
        g[2] = ...
        ...
    end

and

    function v(t, q, p, v)
        v[1] = ...
        v[2] = ...
        ...
    end

Constructors

LODE(ϑ, f, ω, v̄, f̄, t₀, q₀, p₀, λ₀, lagrangian, invariants, parameters, periodicity)

LODE(ϑ, f, l, ω, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector=zero(q₀); kwargs...)
LODE(ϑ, f, l, ω, q₀::StateVector, p₀::StateVector, λ₀::StateVector=zero(q₀); kwargs...)
LODE(ϑ, f, l, ω, t₀, q₀::State, p₀::State, λ₀::StateVector=zero(q₀); kwargs...)
LODE(ϑ, f, l, ω, q₀::State, p₀::State, λ₀::StateVector=zero(q₀); kwargs...)

Keyword arguments

v̄ = (t,q,v) -> nothing

f̄ = f

invariants = nothing

parameters = nothing

periodicity = nothing

GeometricEquations.LODEProblem — Type

LODEProblem: Lagrangian Ordinary Differential Equation Problem

Defines an implicit initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= ϑ(t, q(t), v(t)) \end{aligned}\]

with momentum $p$ and force field $f$, given by

\[\begin{aligned} p &= \frac{\partial L}{\partial v} , & f &= \frac{\partial L}{\partial q} , \end{aligned}\]

and initial conditions $(q_{0}, p_{0})$. This is a special case of a differential algebraic equation with dynamical variables $(q,p)$ and algebraic variable $v$.

GeometricEquations.ODE — Type

ODE: Ordinary Differential Equation

Defines an initial value problem

\[\dot{q} (t) = v(t, q(t)) , \qquad q(t_{0}) = q_{0} ,\]

with vector field $v$, initial condition $q_{0}$ and the solution $q$ taking values in $\mathbb{R}^{d}$.

Parameters

vType <: Callable: type of v

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

v: function computing the vector field

invariants: functions for the computation of invariants, either a NamedTuple containing the equation's invariants or NullInvariants

parameters: type constraints for parameters, either a NamedTuple containing the equation's parameters or NullParameters

periodicity: determines the periodicity of the state vector q for cutting periodic solutions, either a AbstractArray or NullPeriodicity

The function v providing the vector field must have the interface

    function v(t, q, v, params)
        v[1] = ...
        v[2] = ...
        ...
    end

where t is the current time, q is the current solution vector, v is the vector which holds the result of evaluating the vector field $v$ on t and q, and params are additional parameters.

Constructors

ODE(v, invariants, parameters, periodicity)
ODE(v; invariants=NullInvariants(), parameters=NullParameters(), periodicity=NullPeriodicity())

GeometricEquations.ODEProblem — Type

ODEProblem: Ordinary Differential Equation Problem

Defines an initial value problem

\[\dot{q} (t) = v(t, q(t)) , \qquad q(t_{0}) = q_{0} ,\]

with vector field $v$ and initial condition $q_{0}$.

Constructors

ODE(v, tspan, tstep, ics::NamedTuple; kwargs...)
ODE(v, tspan, tstep, q₀::State; kwargs...)

where ics is a NamedTuple with entry q.

Keyword arguments:

invariants = NullInvariants()

parameters = NullParameters()

periodicity = NullPeriodicity()

GeometricEquations.PDAE — Type

PDAE: Partitioned Differential Algebraic Equation

Defines a partitioned differential algebraic initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t, q(t), p(t)) + u(t, q(t), p(t), \lambda(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) + g(t, q(t), p(t), \lambda(t)) , & p(t_{0}) &= p_{0} , \\ 0 &= \phi (t, q(t), p(t), \lambda(t)) , & \lambda(t_{0}) &= \lambda_{0} , \end{aligned}\]

with vector fields $v$ and $f$, projection $u$ and $g$, algebraic constraint $\phi=0$, conditions $(q_{0}, p_{0})$ and $\lambda_{0}$, the dynamical variables $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$ and the algebraic variable $\lambda$ taking values in $\mathbb{R}^{m}$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

vType <: Function: type of v

fType <: Function: type of f

uType <: Function: type of u

gType <: Function: type of g

ϕType <: Function: type of ϕ

ūType <: Function: type of ū

ḡType <: Function: type of ḡ

ψType <: Function: type of ψ

v̄Type <: Function: type of v̄

f̄Type <: Function: type of f̄

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $f$ and $p$

m: dimension of algebraic variable $\lambda$ and the constraint $\phi$

v: function computing the vector field $v$

f: function computing the vector field $f$

u: function computing the projection for $q$

g: function computing the projection for $p$

ϕ: algebraic constraints

ū: function computing the secondary projection field $\bar{u}$ (optional)

ḡ: function computing the secondary projection field $\bar{g}$ (optional)

ψ: secondary constraints (optional)

v̄: function computing an initial guess for the velocity field $v$ (optional)

f̄: function computing an initial guess for the force field $f$ (optional)

t₀: initial time

q₀: initial condition for dynamical variable $q$

p₀: initial condition for dynamical variable $p$

λ₀: initial condition for algebraic variable $\lambda$

μ₀: initial condition for algebraic variable $μ$ (optional)

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

Constructors

PDAE(v, f, u, g, ϕ, ū, ḡ, ψ, v̄, f̄, t₀, q₀, p₀, λ₀, invariants, parameters, periodicity)

PDAE(v, f, u, g, ϕ, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
PDAE(v, f, u, g, ϕ, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
PDAE(v, f, u, g, ϕ, t₀, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
PDAE(v, f, u, g, ϕ, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)

PDAE(v, f, u, g, ϕ, ū, ḡ, ψ, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
PDAE(v, f, u, g, ϕ, ū, ḡ, ψ, q₀::StateVector, p₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
PDAE(v, f, u, g, ϕ, ū, ḡ, ψ, t₀, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
PDAE(v, f, u, g, ϕ, ū, ḡ, ψ, q₀::State, p₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)

GeometricEquations.PDAEProblem — Type

PDAEProblem: Partitioned Differential Algebraic Equation Problem

Defines a partitioned differential algebraic initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t, q(t), p(t)) + u(t, q(t), p(t), \lambda(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) + g(t, q(t), p(t), \lambda(t)) , & p(t_{0}) &= p_{0} , \\ 0 &= \phi (t, q(t), p(t), \lambda(t)) , & \lambda(t_{0}) &= \lambda_{0} , \end{aligned}\]

with vector fields $v$ and $f$, projection $u$ and $g$, algebraic constraint $\phi=0$, and initial conditions $(q_{0}, p_{0})$ and $\lambda_{0}$.

GeometricEquations.PODE — Type

PODE: Partitioned Ordinary Differential Equation

Defines a partitioned initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t, q(t), p(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) , & p(t_{0}) &= p_{0} , \end{aligned}\]

with vector fields $v$ and $f$, initial conditions $(q_{0}, p_{0})$ and the solution $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$.

Parameters

vType <: Function: type of v

fType <: Function: type of f

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

v: function computing the vector field $v$

f: function computing the vector field $f$

invariants: functions for the computation of invariants, either a NamedTuple containing the equation's invariants or NullInvariants

parameters: type constraints for parameters, either a NamedTuple containing the equation's parameters or NullParameters

periodicity: determines the periodicity of the state vector q for cutting periodic solutions, either a AbstractArray or NullPeriodicity

The functions v and f must have the interface

    function v(t, q, p, v, params)
        v[1] = ...
        v[2] = ...
        ...
    end

and

    function f(t, q, p, f, params)
        f[1] = ...
        f[2] = ...
        ...
    end

where t is the current time, q and p are the current solution vectors, v and f are the vectors which hold the result of evaluating the vector fields $v$ and $f$ on t, q and p, and params are additional parameters.

Constructors

PODE(v, f, invariants, parameters, periodicity)
PODE(v, f; invariants=NullInvariants(), parameters=NullParameters(), periodicity=NullPeriodicity())

GeometricEquations.PODEProblem — Type

PODEProblem: Partitioned Ordinary Differential Equation Problem

Defines a partitioned initial value problem

\[\begin{aligned} \dot{q} (t) &= v(t, q(t), p(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) , & p(t_{0}) &= p_{0} , \end{aligned}\]

with vector fields $v$ and $f$ and initial conditions $(q_{0}, p_{0})$.

Constructors

PODE(v, f, tspan, tstep, ics; kwargs...)
PODE(v, f, tspan, tstep, q₀::State, p₀::State; kwargs...)

where ics is a NamedTuple with entries q and p.

Keyword arguments:

invariants = NullInvariants()

parameters = NullParameters()

periodicity = NullPeriodicity()

GeometricEquations.PSDE — Type

PSDE: Stratonovich Partitioned Stochastic Differential Equation

Defines a partitioned stochastic differential initial value problem

\[\begin{aligned} dq (t) &= v(t, q(t)) \, dt + B(t, q(t)) \circ dW , & q(t_{0}) &= q_{0} , \\ dp (t) &= f(t, q(t)) \, dt + G(t, q(t)) \circ dW , & p(t_{0}) &= p_{0} \end{aligned}\]

with the drift vector fields $v$ and $f$, diffusion matrices $B$ and $G$, initial conditions $q_{0}$ and $p_{0}$, the dynamical variables $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$, and the m-dimensional Wiener process W

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

vType <: Function: type of v

fType <: Function: type of f

BType <: Function: type of B

GType <: Function: type of G

pType <: Union{NamedTuple,Nothing}: parameters type

Fields

d: dimension of dynamical variable $q$ and the vector field $v$

m: dimension of the Wiener process

ns: number of sample paths

v: function computing the drift vector field for the position variable $q$

f: function computing the drift vector field for the momentum variable $p$

B: function computing the d x m diffusion matrix for the position variable $q$

G: function computing the d x m diffusion matrix for the momentum variable $p$

t₀: initial time

q₀: initial condition for dynamical variable $q$ (may be a random variable itself)

p₀: initial condition for dynamical variable $p$ (may be a random variable itself)

parameters: either a NamedTuple containing the equations parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

The functions v, f, B and G, providing the drift vector fields and diffusion matrices, take four arguments, v(t, q, p, v), f(t, q, p, f), B(t, q, p, B) and G(t, q, p, G), where t is the current time, (q, p) is the current solution vector, and v, f, B and G are the variables which hold the result of evaluating the vector fields $v$, $f$ and the matrices $B$, $G$ on t and (q,p).

Constructors

PSDE(m, ns, v, f, B, G, t₀, q₀, p₀; parameters=NullParameters(), periodicity=zero(q₀[begin]))
PSDE(m, ns, v, f, B, G, q₀::StateVector, p₀::StateVector; kwargs...) = PSDE(m, ns, v, f, B, G, 0.0, q₀, p₀; kwargs...)
PSDE(m, ns, v, f, B, G, t₀, q₀::State, p₀::State; kwargs...) = PSDE(m, ns, v, f, B, G, t₀, [q₀], [p₀]; kwargs...)
PSDE(m, ns, v, f, B, G, q₀::State, p₀::State; kwargs...) = PSDE(m, ns, v, f, B, G, 0.0, q₀, p₀; kwargs...)

Example

    function v(λ, t, q, v)
        v[1] = λ*q[1]
        v[2] = λ*q[2]
    end

    function B(μ, t, q, B)
        B[1] = μ*q[1]
        B[2] = μ*q[2]
    end

    t₀ = 0.
    q₀ = [1., 1.]
    λ  = 2.
    μ  = 1.

    v_sde = (t, q, v) -> v(λ, t, q, v)
    B_sde = (t, q, B) -> B(μ, t, q, B)

    sde = SDE(v_sde, B_sde, t₀, q₀)

GeometricEquations.PSDEProblem — Type

GeometricEquations.SDE — Type

SDE: Stratonovich Stochastic Differential Equation

Defines a stochastic differential initial value problem

\[\begin{aligned} dq (t) &= v(t, q(t)) \, dt + B(t, q(t)) \circ dW , & q(t_{0}) &= q_{0} , \end{aligned}\]

with drift vector field $v$, diffusion matrix $B$, initial conditions $q_{0}$, the dynamical variable $q$ taking values in $\mathbb{R}^{d}$, and the m-dimensional Wiener process W

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

vType <: Function: type of v

BType <: Function: type of B

pType <: Union{NamedTuple,Nothing}: parameters type

Fields

d: dimension of dynamical variable $q$ and the vector field $v$

m: dimension of the Wiener process

ns: number of sample paths

v: function computing the deterministic vector field

B: function computing the d x m diffusion matrix

t₀: initial time

q₀: initial condition for dynamical variable $q$ (may be a random variable itself)

parameters: either a NamedTuple containing the equations parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

The functions v and B, providing the drift vector field and diffusion matrix, v(t, q, v) and B(t, q, B, col=0), where t is the current time, q is the current solution vector, and v and B are the variables which hold the result of evaluating the vector field $v$ and the matrix $B$ on t and q (if col==0), or the column col of the matrix B (if col>0).

Constructors

SDE(m, ns, v, B, t₀, q₀; parameters=NullParameters(), periodicity=zero(q₀[begin]))
SDE(m, ns, v, B, q₀::StateVector; kwargs...) = SDE(m, ns, v, B, 0.0, q₀; kwargs...)
SDE(m, ns, v, B, t₀, q₀::State; kwargs...) = SDE(m, ns, v, B, t₀, [q₀]; kwargs...)
SDE(m, ns, v, B, q₀::State; kwargs...) = SDE(m, ns, v, B, 0.0, q₀; kwargs...)

Example

    function v(t, q, v, p)
        λ = p[:λ]
        v[1] = λ*q[1]
        v[2] = λ*q[2]
    end

    function B(t, q, B, p, col=0)
        μ = p[:μ]
        if col==0 #whole matrix
            B[1,1] = μ*q[1]
            B[2,1] = μ*q[2]
        elseif col==1
            #just first column
        end
    end

    t₀ = 0.
    q₀ = [1., 1.]
    λ  = 2.
    μ  = 1.
    p = (λ=λ, μ=μ)

    sde = SDE(v, B, t₀, q₀; parameters=p)

GeometricEquations.SDEProblem — Type

GeometricEquations.SODE — Type

SODE: Split Ordinary Differential Equation

Defines an initial value problem

\[\dot{q} (t) = v(t, q(t)) , \qquad q(t_{0}) = q_{0} ,\]

with vector field $v$, initial condition $q_{0}$ and the solution $q$ taking values in $\mathbb{R}^{d}$. Here, the vector field $v$ is given as a sum of vector fields

\[v (t) = v_1 (t) + ... + v_r (t) .\]

Parameters

vType <: Union{Tuple,Nothing}: type of v

qType <: Union{Tuple,Nothing}: type of q

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

v: tuple of functions computing the vector field

q: tuple of functions computing the solution

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

The functions v_i providing the vector field must have the interface

    function v_i(t, q, v, params)
        v[1] = ...
        v[2] = ...
        ...
    end

and the functions q_i providing the solutions must have the interface

    function q_i(t, q₀, q₁, h, params)
        q₁[1] = q₀[1] + ...
        q₁[2] = q₀[2] + ...
        ...
    end

where t is the current time, q₀ is the current solution vector, q₁ is the new solution vector which holds the result of computing one substep with the vector field $v_i$ on t and q₀, and h is the (sub-)timestep to compute the update for.

Constructors

SODE(v, invariants, parameters, periodicity)
SODE(v; invariants=NullInvariants(), parameters=NullParameters(), periodicity=NullPeriodicity())
SODE(v, q, invariants, parameters, periodicity)
SODE(v, q; invariants=NullInvariants(), parameters=NullParameters(), periodicity=NullPeriodicity())

GeometricEquations.SODEProblem — Type

SODEProblem: Split Ordinary Differential Equation Problem

Defines an initial value problem

\[\dot{q} (t) = v(t, q(t)) , \qquad q(t_{0}) = q_{0} ,\]

with vector field $v$ and initial condition $q_{0}$. Here, the vector field $v$ is given as a sum of vector fields

\[v (t) = v_1 (t) + ... + v_r (t) .\]

Fields

equation: reference to the parent equation object

functions: all functions (e.g. vector fields) provided by the equation

tspan: time span for problem (t₀,t₁)

tstep: time step to be used in simulation

ics: initial condition (NamedTuple containing the field q)

parameters: either a NamedTuple containing the equation's parameters or nothing

The functions v_i providing the vector field must have the interface

    function v_i(t, q, v, params)
        v[1] = ...
        v[2] = ...
        ...
    end

and the functions q_i providing the solutions must have the interface

    function q_i(t, q₀, q₁, h, params))
        q₁[1] = q₀[1] + ...
        q₁[2] = q₀[2] + ...
        ...
    end

where t is the current time, q₀ is the current solution vector, q₁ is the new solution vector which holds the result of computing one substep with the vector field $v_i$ on t and q₀, and h is the (sub-)timestep to compute the update for.

The fact that the function v returns the solution and not just the vector field for each substep increases the flexibility for the use of splitting methods, e.g., it allows to use another integrator for solving substeps.

Constructors

SODE(v, q, t₀, q₀, invariants, parameters, periodicity)

SODE(v, q::Union{Tuple,Nothing}, t₀::Real, q₀::StateVector; kwargs...)
SODE(v, q::Union{Tuple,Nothing}, t₀::Real, q₀::State; kwargs...)
SODE(v, q::Union{Tuple,Nothing}, q₀::StateVector; kwargs...)
SODE(v, q::Union{Tuple,Nothing}, q₀::State; kwargs...)

SODE(v, t₀::Real, q₀::StateVector; kwargs...)
SODE(v, t₀::Real, q₀::State; kwargs...)
SODE(v, q₀::StateVector; kwargs...)
SODE(v, q₀::State; kwargs...)

GeometricEquations.SPDAE — Type

SPDAE: Split Partitioned Differential Algebraic Equation EXPERIMENTAL

Defines a split differential algebraic initial value problem, that is a canonical Hamiltonian system of equations subject to Dirac constraints,

\[\begin{aligned} \dot{q} (t) &= v_1(t, q(t), p(t)) + v_2(t, q(t), p(t), \lambda(t)) + v_3(t, q(t), p(t), \lambda(t), \gamma(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f_1(t, q(t), p(t)) + f_2(t, q(t), p(t), \lambda(t)) + f_3(t, q(t), p(t), \lambda(t), \gamma(t)) , & p(t_{0}) &= p_{0} , \\ 0 &= \phi (t, q(t), p(t)) , \\ 0 &= \psi (t, q(t), p(t), \dot{q}(t), \dot{p}(t)) , \end{aligned}\]

with vector fields $v_i$ and $f_i$ for $i = 1 ... 3$, primary constraint $\phi(q,p)=0$ and secondary constraint $\psi(q,p,\lambda)=0$, initial conditions $(q_{0}, p_{0})$, the dynamical variables $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$ and the algebraic variables $(\lambda, \gamma)$ taking values in $\mathbb{R}^{n} \times \mathbb{R}^{d}$.

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

vType <: Function: type of v

fType <: Function: type of f

ϕType <: Function: type of ϕ

ψType <: Function: type of ψ

invType <: OptionalNamedTuple: invariants type

parType <: OptionalNamedTuple: parameters type

perType <: OptionalArray{AT}: periodicity type

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $v$ and $f$

m: dimension of algebraic variables $\lambda$ and $\gamma$ and the constraints $\phi$ and $\psi$

v: tuple of functions computing the vector fields $v_i$, $i = 1 ... 3$

f: tuple of functions computing the vector fields $f_i$, $i = 1 ... 3$

ϕ: primary constraints

ψ: secondary constraints

t₀: initial time

q₀: initial condition for dynamical variable $q$

p₀: initial condition for dynamical variable $p$

λ₀: initial condition for algebraic variable $λ$

μ₀: initial condition for algebraic variable $μ$ (optional)

invariants: either a NamedTuple containing the equation's invariants or nothing

parameters: either a NamedTuple containing the equation's parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

Constructors

SPDAE(v, f, ϕ, ψ, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...)
SPDAE(v, f, ϕ, ψ, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...)
SPDAE(v, f, ϕ, ψ, t₀, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...)
SPDAE(v, f, ϕ, ψ, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...)

Keyword arguments:

invariants = nothing

parameters = nothing

periodicity = nothing

GeometricEquations.SPDAEProblem — Type

GeometricEquations.SPSDE — Type

SPSDE: Stratonovich Split Partitioned Stochastic Differential Equation

Defines a partitioned stochastic differential initial value problem

\[\begin{aligned} dq (t) &= v(t, q(t)) \, dt + B(t, q(t)) \circ dW , & q(t_{0}) &= q_{0} , \\ dp (t) &= [ f_1(t, q(t)) + f_2(t, q(t)) ] \, dt + [ G_1(t, q(t)) + G_2(t, q(t)) ] \circ dW , & p(t_{0}) &= p_{0} , \end{aligned}\]

with the drift vector fields $v$ and $f_i$, diffusion matrices $B$ and $G_i$, initial conditions $q_{0}$ and $p_{0}$, the dynamical variables $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$, and the m-dimensional Wiener process W

Parameters

DT <: Number: data type

TT <: Real: time step type

AT <: AbstractArray{DT}: array type

vType <: Function: type of v

f1Type <: Function: type of f1

f2Type <: Function: type of f2

BType <: Function: type of B

G1Type <: Function: type of G1

G2Type <: Function: type of G2

pType <: Union{NamedTuple,Nothing}: parameters type

Fields

d: dimension of dynamical variable $q$ and the vector fields $vi$

m: dimension of the Wiener process

ni: number of initial conditions

ns: number of sample paths

v : function computing the drift vector field for the position variable $q$

f1: function computing the drift vector field for the momentum variable $p$

f2: function computing the drift vector field for the momentum variable $p$

B : function computing the d x m diffusion matrix for the position variable $q$

G1: function computing the d x m diffusion matrix for the momentum variable $p$

G2: function computing the d x m diffusion matrix for the momentum variable $p$

t₀: initial time

q₀: initial condition for dynamical variable $q$ (may be a random variable itself)

p₀: initial condition for dynamical variable $p$ (may be a random variable itself)

parameters: either a NamedTuple containing the equations parameters or nothing

periodicity: determines the periodicity of the state vector q for cutting periodic solutions

The functions v, f, B and G, providing the drift vector fields and diffusion matrices, take four arguments, v(t, q, p, v), f(t, q, p, f), B(t, q, p, B) and G(t, q, p, G), where t is the current time, (q, p) is the current solution vector, and v, f, B and G are the variables which hold the result of evaluating the vector fields $v$, $f$ and the matrices $B$, $G$ on t and (q,p).

Constructors

SPSDE(m, ns, v, f1, f2, B, G1, G2, t₀, q₀, p₀; parameters=NullParameters(), periodicity=zero(q₀[begin]))
SPSDE(m, ns, v, f1, f2, B, G1, G2, q₀::StateVector, p₀::StateVector; kwargs...) = SPSDE(m, ns, v, f1, f2, B, G1, G2, 0.0, q₀, p₀; kwargs...)
SPSDE(m, ns, v, f1, f2, B, G1, G2, t₀, q₀::State, p₀::State; kwargs...) = SPSDE(m, ns, v, f1, f2, B, G1, G2, t₀, [q₀], [p₀]; kwargs...)
SPSDE(m, ns, v, f1, f2, B, G1, G2, q₀::State, p₀::State; kwargs...) = SPSDE(m, ns, v, f1, f2, B, G1, G2, 0.0, q₀, p₀; kwargs...)

Example

    function v(λ, t, q, v)
        v[1] = λ*q[1]
        v[2] = λ*q[2]
    end

    function B(μ, t, q, B)
        B[1] = μ*q[1]
        B[2] = μ*q[2]
    end

    t₀ = 0.
    q₀ = [1., 1.]
    λ  = 2.
    μ  = 1.

    v_sde = (t, q, v) -> v(λ, t, q, v)
    B_sde = (t, q, B) -> B(μ, t, q, B)

    sde = SDE(v_sde, B_sde, t₀, q₀)

GeometricEquations.SPSDEProblem — Type

GeometricBase.equation — Method

Returns the parent equation object of the problem.

GeometricBase.functions — Method

Returns a NamedTuple containing all functions (e.g. vector fields) provided by the equation.

GeometricBase.invariants — Method

Returns a NamedTuple containing all invariants provided by the equation.

GeometricBase.solutions — Method

Returns a NamedTuple containing all solutions provided by the equation.

GeometricEquations.promote_tspan — Method

promote_tspan(tspan)

Convert the tspan field of a GeometricProblem to a (tmin, tmax) tuple, where both elements are of the same type.